+++ test/support/wrapper.ex
@@ -1,0 +1,5 @@
+defmodule NumberWrapper do
+  def wrapper(string, type) do
+    {:ok, "<#{type}>" <> string <> "<#{type}>"}
+  end
+end